requirements.txt
Flask==0.10.1
Jinja2==2.7.3
MarkupSafe==0.23
Werkzeug==0.9.6
google-api-python-client==1.2
httplib2==0.9
itsdangerous==0.24
python-gflags==2.0
wsgiref==0.1.2
